fix: scope enum defined in the wrong module
