+++ go-sdk/sdk/centralized-configuration/centralized_configuration.go
@@ -20,6 +20,16 @@ const (
 	_centralizedConfigurationLoggerName = "[CENTRALIZED CONFIGURATION]"
 )
 
+type Scope string
+
+const (
+	GlobalScope          Scope = "global"
+	ProductScope         Scope = "product"
+	WorkflowScope        Scope = "workflow"
+	ProcessScope         Scope = "process"
+	_messagingLoggerName       = "[MESSAGING]"
+)
+
 type CentralizedConfiguration struct {
 	logger     logr.Logger
 	globalKv   nats.KeyValue
